changed PlayerServer to using a MapSet as datastructure
Currently the read just return a MapSet.to_list on the state,
causing the rest of the application to behave the same. This fixes
our only pending test. PlayerServer.add may now be desirable to return
whether or not the player was added, to indicate whether it is a
duplicate, such that we can respond appropriately.