integrated no duplicate summoners into players_channel and tests
This ended up requiring using a Map instead of a MapSet, as Player
structs would differ due to getting different id's. As such we're not
using their summoner name as the Key in PlayerServers. This requires
that all PlayerServers represent a single region, as we can run into
name conflicts otherwise. Since this is already the case, it's not a
problem, but we may want to make this summoner_id eventually, and thus
store this in the Player as well.