made public unmatch/match broadcast functions in PlayersChannel.
The unmatch broadcast is now used from the PlayerServer, both to
simplify its code, as well as to remove the need for PlayerServers to be
aware of what events indicate what etc. This has the advantage of
reducing the function nesting in PlayerServers handle_info for leaves by
1. We're still missing 2 credo refactor issues, both of which can be
handled temporarily by removing the lazy logging call. A more stable
solution is however needed.