auto revision for upgrades and added excoveralls
Additionally add excoveralls to travis. The auto revisioning for each
app is what allows us to smoothly due hot code upgrades on prod.
Currently however, this causes pre_commits to fail, as the System.cmd is
run from a non-git directory causing System.cmd("git") to result in
error. This is probably a current limitation to pre_commit_hook for
elixir that we'll have to look at, as we definitely still want to have
these running.