intial working access_server for token handling.
Refresh token does not work yet, and it fails when ran in test,
as it in its init function currently depends on environment variables,
as it already then tries to fetch an access token.

Should probably just instantiate it with dummy state, where the
expiration date, is the small DateTime, such that first call to the
server triggers the token fetch. Yet to be discovered why refresh token
fails.

May also be beneficial to add a task for pinging the AccessServer with some
certain interval, as the refresh_token can expire otherwise.